rename hand module to players
